test(onboard): split VM DNS monkeypatch coverage
